Added trainer and Traefik
